Correct "UTF-16" comment in IXMLWriter.h Fix cast warning in CTerrainSceneNode.cpp Fix include paths in MSVC test projects. Reduce the number of test iterations in fast_atof.cpp test; the duration delta is still clear. Remove screenshots.cpp test, since drawPixel.cpp (and other future tests) already do more.
Test suite pass at GMT Sat Nov 29 08:12:52 2008

git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/releases/1.5@1867 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CTerrainSceneNode.cpp
@@ -352,7 +352,7 @@ namespace scene
 							mb->drop();
 							return false;
 						}
-						vertex.Pos.Y=val;
+						vertex.Pos.Y=(f32)val;
 					}
 					break;
 				}
